requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
Django==2.2.10
django-cors-headers==3.3.0
django-extensions==2.2.9
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.2
pytest-cov==2.9.0
pytest-django==3.9.0
python-coveralls==2.9.3
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
